Make panic notif always appear
